# Refactor rarely changing engine state into its own struct