--- a/crates/nu-command/src/commands/classified/external.rs
+++ b/crates/nu-command/src/commands/classified/external.rs
@@ -27,12 +27,7 @@ pub(crate) fn run_external_command(
     trace!(target: "nu::run::external", "-> {}", command.name);
 
     context.sync_path_to_env();
-    if !context
-        .engine_state
-        .host
-        .lock()
-        .is_external_cmd(&command.name)
-    {
+    if !context.host().lock().is_external_cmd(&command.name) {
         return Err(ShellError::labeled_error(
             "Command not found",
             format!("command {} not found", &command.name),
@@ -64,7 +59,7 @@ fn run_with_stdin(
     input: InputStream,
     external_redirection: ExternalRedirection,
 ) -> Result<InputStream, ShellError> {
-    let path = context.engine_state.shell_manager.path();
+    let path = context.shell_manager().path();
 
     let mut command_args = vec![];
     for arg in command.args.iter() {