--- a/crates/nu-command/src/commands/run_external.rs
+++ b/crates/nu-command/src/commands/run_external.rs
@@ -148,7 +148,7 @@ fn maybe_autocd_dir(cmd: &ExternalCommand, ctx: &mut EvaluationContext) -> Optio
             if name.ends_with(':') {
                 // This looks like a drive shortcut. We need to a) switch drives and b) go back to the previous directory we were viewing on that drive
                 // But first, we need to save where we are now
-                let current_path = ctx.shell_manager.path();
+                let current_path = ctx.shell_manager().path();
 
                 let split_path: Vec<_> = current_path.split(':').collect();
                 if split_path.len() > 1 {