This moves the rarely changing engine state into its own struct, which allows us to pay less for each clone. I've added some convenience methods to access the various parts.

This has the nice side effect of making some of our benchmarks faster. Here is the difference with gradient.nu:

## Before
```
───┬───────────────────
 # │     real time     
───┼───────────────────
 0 │ 169ms 709us 822ns 
───┴───────────────────
```

## After
```
───┬───────────────────
 # │     real time     
───┼───────────────────
 0 │ 137ms 473us 687ns 
───┴───────────────────
```